massive refactoring; Daos don't depend on Services
