+++ src/main/java/cf/baradist/JettyServer.java
@@ -2,44 +2,34 @@ package cf.baradist;
 
 import cf.baradist.controller.EntryPoint;
 import cf.baradist.controller.UserController;
-import cf.baradist.dao.DaoFactory;
-import cf.baradist.dao.H2DaoFactory;
-import cf.baradist.dao.UserDao;
-import cf.baradist.dao.UserDaoImpl;
+import cf.baradist.dao.h2.UserDaoImpl;
 import cf.baradist.service.UserService;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.h2.jdbcx.JdbcDataSource;
+import org.h2.tools.RunScript;
 
-import javax.sql.DataSource;
-
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.sql.Connection;
 import java.sql.SQLException;
-
-import static cf.baradist.dao.H2DaoFactory.H2_CONNECTION_URL;
 
 public class JettyServer {
     public static void main(String[] args) throws Exception {
-        initDb();
+        configureDb();
         runServer();
     }
 
-    private static void initDb() {
+    private static void configureDb() {
         JdbcDataSource ds = new JdbcDataSource();
+        // TODO: move to properties
         ds.setURL("jdbc:h2:mem:moneyapp;DB_CLOSE_DELAY=-1");
         ds.setUser("sa");
         ds.setPassword("sa");
-        try {
-            H2DaoFactory.setConnection(ds.getConnection());
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        DaoFactory daoFactory = DaoFactory.getDaoFactory();
-        daoFactory.fillTestData();
-//        final UserDao userDao__ = DaoFactory.getDaoFactory().getUserDao();
         UserService.getInstance().setUserDao((UserDaoImpl) ds::getConnection);
-//        DaoFactory daoFactory = DaoFactory.getDaoFactory();
-//        daoFactory.fillTestData();
+
+        fillTestData(ds);
     }
 
     private static void runServer() throws Exception {
@@ -64,6 +54,16 @@ public class JettyServer {
             jettyServer.join();
         } finally {
             jettyServer.destroy();
+        }
+    }
+
+    public static void fillTestData(JdbcDataSource ds) {
+        try (Connection conn = ds.getConnection()) {
+            RunScript.execute(conn, new FileReader("src/main/resources/init.sql"));
+        } catch (SQLException e) {
+            throw new RuntimeException("fillTestData(): ", e);
+        } catch (FileNotFoundException e) {
+            throw new RuntimeException("fillTestData(): can't find a file \"src/main/resources/init.sql\"", e);
         }
     }
 }
